fix(plone resources): disable resources we don't want from Plone
Plone is full of resources which add features and styles.
The best way to disabled resources is not to disabled them from
portal_resources but to make them empty in a skin directory.
This is a workaround to make it easier to uninstall the theme
but also to create new theme based on this one.